Moved `Document#getNearesetSelectionRange` to `Schema#getNearesetSelectionRange`.
